Remove locale-specific files and simplify planner
