+++ middleware.ts
@@ -1,0 +1,11 @@
+import createMiddleware from 'next-intl/middleware';
+import { locales, defaultLocale } from './i18n';
+
+export default createMiddleware({
+  locales,
+  defaultLocale
+});
+
+export const config = {
+  matcher: ['/((?!api|_next|favicon.ico|sw.js|manifest.json).*)']
+};
