+++ next-intl.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  locales: ['en', 'es'],
+  defaultLocale: 'es',
+  localeDetection: false
+};
