+++ public/manifest.json
@@ -1,0 +1,12 @@
+{
+  "name": "VisitAtlantico",
+  "short_name": "VisitAtlantico",
+  "start_url": "/",
+  "display": "standalone",
+  "background_color": "#ffffff",
+  "theme_color": "#006994",
+  "icons": [
+    { "src": "/logo.png", "sizes": "192x192", "type": "image/png" },
+    { "src": "/logo.png", "sizes": "512x512", "type": "image/png" }
+  ]
+}
