+++ public/sw.js
@@ -1,0 +1,10 @@
+const CACHE_NAME = 'va-pwa-v1';
+const urls = ['/','/planner'];
+self.addEventListener('install', e => {
+  e.waitUntil(caches.open(CACHE_NAME).then(c => c.addAll(urls)));
+});
+self.addEventListener('fetch', e => {
+  e.respondWith(
+    caches.match(e.request).then(r => r || fetch(e.request))
+  );
+});
