+++ src/components/ItineraryStopCard.tsx
@@ -19,6 +19,7 @@ export interface Stop {
   name: string;
   lat: number;
   lng: number;
+  day: number;
   startTime: string;
   durationMinutes: number;
   description: string;
